backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
sqlmodel
requests
python-multipart
python-dotenv
google-auth
PyJWT
